HOCS-2213: add sanity checks for house addresses
At present when trying to refresh the member list, 
if the house address for the member does not exist 
this causes a Null Pointer exception later. 

This change adds exception handling for if this 
entity does not exist. This logs the error and then 
returns an empty set. This allows for the refresh to 
continue past this point and potentially update other 
member lists. To assist with removing magic strings, 
a string enum has been added for the house codes.